package strategy classes and bridge classes and refactor accordingly
