+++ src/main/java/com/synopsys/integration/jenkins/scan/ScanPipelineCommands.java
@@ -4,9 +4,9 @@ import com.synopsys.integration.jenkins.scan.bridge.BridgeDownloadManager;
 import com.synopsys.integration.jenkins.scan.bridge.BridgeDownloadParameters;
 import com.synopsys.integration.jenkins.scan.exception.ScannerJenkinsException;
 import com.synopsys.integration.jenkins.scan.global.ExceptionMessages;
-import com.synopsys.integration.jenkins.scan.service.BridgeDownloadParametersService;
-import com.synopsys.integration.jenkins.scan.service.scan.ScanStrategyFactory;
-import com.synopsys.integration.jenkins.scan.service.scan.ScanStrategyService;
+import com.synopsys.integration.jenkins.scan.service.bridge.BridgeDownloadParametersService;
+import com.synopsys.integration.jenkins.scan.service.scan.strategy.ScanStrategyFactory;
+import com.synopsys.integration.jenkins.scan.service.scan.strategy.ScanStrategy;
 import hudson.FilePath;
 import hudson.model.TaskListener;
 import java.util.Map;
@@ -23,7 +23,7 @@ public class ScanPipelineCommands {
     }
 
     public int runScanner(Map<String, Object> scanParameters, ScanStrategyFactory scanStrategyFactory) throws ScannerJenkinsException {
-        ScanStrategyService scanStrategyService = scanStrategyFactory.getParametersService(scanParameters);
+        ScanStrategy scanStrategy = scanStrategyFactory.getParametersService(scanParameters);
 
         BridgeDownloadParameters bridgeDownloadParameters = new BridgeDownloadParameters(workspace, listener);
         BridgeDownloadParametersService bridgeDownloadParametersService = new BridgeDownloadParametersService(workspace, listener);
@@ -31,7 +31,7 @@ public class ScanPipelineCommands {
 
         int exitCode = -1;
 
-        if (scanStrategyService.isValidScanParameters(scanParameters) &&
+        if (scanStrategy.isValidScanParameters(scanParameters) &&
             bridgeDownloadParametersService.performBridgeDownloadParameterValidation(bridgeDownloadParams)) {
             BridgeDownloadManager bridgeDownloadManager = new BridgeDownloadManager(workspace, listener);
             boolean isBridgeDownloadRequired = bridgeDownloadManager.isSynopsysBridgeDownloadRequired(bridgeDownloadParams);
@@ -44,7 +44,7 @@ public class ScanPipelineCommands {
             FilePath bridgeInstallationPath = new FilePath(workspace.getChannel(), bridgeDownloadParams.getBridgeInstallationPath());
 
             try {
-                exitCode = scanner.runScanner(scanParameters, scanStrategyService, bridgeInstallationPath);
+                exitCode = scanner.runScanner(scanParameters, scanStrategy, bridgeInstallationPath);
             } catch (Exception e) {
                 e.printStackTrace(listener.getLogger());
                 throw new ScannerJenkinsException(ExceptionMessages.scannerFailureMessage(e.getMessage()));
