+++ src/main/java/com/synopsys/integration/jenkins/scan/service/scan/strategy/ScanStrategy.java
@@ -1,9 +1,9 @@
-package com.synopsys.integration.jenkins.scan.service.scan;
+package com.synopsys.integration.jenkins.scan.service.scan.strategy;
 
 import com.synopsys.integration.jenkins.scan.global.enums.ScanType;
 import java.util.Map;
 
-public interface ScanStrategyService {
+public interface ScanStrategy {
     ScanType getScanType();
 
     boolean isValidScanParameters(Map<String, Object> parametersMap);
